+++ src/codecool/com/subway/orderitem/OrderItem.java
@@ -1,5 +1,5 @@
 package codecool.com.subway.orderitem;
 
 public abstract class OrderItem {
-    public abstract int calculatePrice();
+    public abstract double calculatePrice();
 }
